Properly discover existing image in OCI target (#181)
When using ghcr.io as an OCI target then charts-syncer would always
overwrite an already existing chart because ghcr.io requires a client
to retrieve a bearer token before making any requests against it. This
is also true when basic auth credentials are provided. The `Repo#Has`
method would always fail in that case because it can't handle 401
responses.

I exchanged the code in the `Has` method to use the
go-containerregistry client instead that properly handles
authentication.

Signed-off-by: Max Jonas Werner <mail@makk.es>